author unfollow local users
